requirements.txt
flask
flask-cors
gunicorn 
psycopg2-binary
werkzeug
